Module hexchat_api::callback_data

The CallbackData object holds all the information about a callback needed to manage the user_data and invoke it safely and correctly. The objects of this module are used internally. This file also contains type declarations for the Rust-facing callback signatures.

Re-exports

use std::any::Any;

use std::mem::ManuallyDrop;

use crate::hook::Hook;

use crate::hexchat::Hexchat;

use crate::hexchat::Eat;

use crate::hexchat::EventAttrs;

use crate::utils::*;

Structs

CallbackData
Pointers to instances of this struct are registered with the Hexchat callbacks. On the C-facing side, this is the user_data passed to the native callback wrappers (like c_print_callback()). When invoked by Hexchat, the native callbacks receive a pointer to a user_data (CallbackData) object, which the wrapper then uses to invoke the Rust-implemented callback held in the UCallback field below. The data field holds the user data registered for the Rust-facing callback, and is passed to it when invoked.

Enums

CBType
An enumeration of the different types of callback.

Type Definitions

Callback
The Rust-facing function signature corresponding to the C-facing
C_Callback. Note that, unlike the C API, the Rust-facing callback signatures include a reference to the Hexchat pointer for convenience.

FdCallback
The Rust-facing function signature corresponding to the C-facing
C_FdCallback. Note that, unlike the C API, the Rust-facing callback signatures include a reference to the Hexchat pointer for convenience.

MD
Holds the Rust-implemented function, or closure, of a registered Hexchat callback. ManuallyDrop had to be applied to the union's fields to get it to compile - it's a 0-cost abstraction, so no big deal.

PrintAttrsCallback
The Rust-facing function signature corresponding to the C-facing
C_PrintAttrsCallback. Note that, unlike the C API, the Rust-facing callback signatures include a reference to the Hexchat pointer for convenience.

PrintCallback
The Rust-facing function signature corresponding to the C-facing
C_PrintCallback. Note that, unlike the C API, the Rust-facing callback signatures include a reference to the Hexchat pointer for convenience.

TimerCallback
The Rust-facing function signature corresponding to the C-facing
C_TimerCallback. Note that, unlike the C API, the Rust-facing callback signatures include a reference to the Hexchat pointer for convenience.

UserData
The Rust-facing user_data type.

Unions

UCallback