Struct hexchat_api::callback_data::CallbackData

pub(crate) struct CallbackData {
    cbtype: CBType,
    callback: UCallback,
    data: Option<Box<dyn Any>>,
    hook: Hook,
}

Pointers to instances of this struct are registered with the Hexchat callbacks. On the C-facing side, this is the user_data passed to the native callback wrappers (like c_print_callback()). When invoked by Hexchat, the native callbacks receive a pointer to a user_data (CallbackData) object, which the wrapper then uses to invoke the Rust-implemented callback held in the UCallback field below. The data field holds the user data registered for the Rust-facing callback, and is passed to it when invoked.

Fields
cbtype: CBTypecallback: UCallbackdata: Option<Box<dyn Any>>hook: Hook
Implementations

impl CallbackData

pub(crate) fn new_command_data(
callback: Box<dyn FnMut(&Hexchat, &[String], &[String], &mut Option<Box<dyn Any>>) -> Eat>,
data: Option<Box<dyn Any>>,
hook: Hook
) -> Self

Creates callback data for a regular command or server command.

pub(crate) fn new_print_data(
callback: Box<dyn FnMut(&Hexchat, &[String], &mut Option<Box<dyn Any>>) -> Eat>,
data: Option<Box<dyn Any>>,
hook: Hook
) -> Self

Creates callback data for a print callback.

pub(crate) fn new_print_attrs_data(
callback: Box<dyn FnMut(&Hexchat, &[String], &EventAttrs, &mut Option<Box<dyn Any>>) -> Eat>,
data: Option<Box<dyn Any>>,
hook: Hook
) -> Self

Creates callback data for a print attrs callback.

pub(crate) fn new_timer_data(
callback: Box<dyn FnMut(&Hexchat, &mut Option<Box<dyn Any>>) -> i32>,
data: Option<Box<dyn Any>>,
hook: Hook
) -> Self

Creates callback data for a timer callback.

pub(crate) fn new_fd_data(
callback: Box<dyn FnMut(&Hexchat, i32, i32, &mut Option<Box<dyn Any>>) -> Eat>,
data: Option<Box<dyn Any>>,
hook: Hook
) -> Self

Creates callback data for a fd callback.

pub(crate) fn get_data(&mut self) -> &mut Option<Box<dyn Any>>

Returns a mutable reference to the Rust-facing user_data that was registered with the callback.

pub(crate) fn take_data(self) -> Option<Box<dyn Any>>

Returns the data (Rust facing user_data) field of the object. Ownership of the user_data is transferred to the caller from this operation. This is used internally by Hook::unhook() to retrieve callback data when a callback is unregistered. This gives the runtime the opportunity to free the data by going out of scope, or perform any custom cleanup. After this function is called, the callback should be considered finished.

pub(crate) unsafe fn command_cb(
&mut self,
hc: &Hexchat,
word: &[String],
word_eol: &[String],
ud: &mut Option<Box<dyn Any>>
) -> Eat

Invokes the callback held in the callback field.

pub(crate) unsafe fn print_cb(
&mut self,
hc: &Hexchat,
word: &[String],
ud: &mut Option<Box<dyn Any>>
) -> Eat

Invokes the callback held in the callback field. This is invoked by c_print_callback() which is the C-side registered callback for each print callback.

pub(crate) unsafe fn print_attrs_cb(
&mut self,
hc: &Hexchat,
word: &[String],
attrs: &EventAttrs,
ud: &mut Option<Box<dyn Any>>
) -> Eat

Invokes the callback held in the callback field. This is invoked by c_print_attrs_callback().

pub(crate) unsafe fn timer_cb(
&mut self,
hc: &Hexchat,
ud: &mut Option<Box<dyn Any>>
) -> i32

Invokes the callback held in the callback field. This is invoked by c_timer_callback()`.

pub(crate) unsafe fn fd_cb(
&mut self,
hc: &Hexchat,
fd: i32,
flags: i32,
ud: &mut Option<Box<dyn Any>>
) -> Eat

Invokes the callback held in the callback field. This is invoked by c_fd_callback()`.