Module hexchat_api::hexchat

This file holds the Hexchat struct used to interface with Hexchat. Its fields are the actual callbacks provided by Hexchat. When Hexchat loads this library, the Hexchat pointer is stored and used by casting it to the struct contained in this file. These native function pointers are private to this file, and a more Rust-friendly API is provided through this Hexchat interface.

Re-exports

use libc::c_int;

use libc::c_char;

use libc::c_void;

use libc::time_t;

use std::any::Any;

use std::ffi::CString;

use std::ffi::CStr;

use std::ops::FnMut;

use std::ptr::null;

use std::str;

use crate::callback_data::CallbackData;

use crate::context::Context;

use crate::hexchat_callbacks::*;

use crate::hook::Hook;

use crate::list_iterator::ListIterator;

use crate::plugin::Plugin;

use crate::utils::*;

use crate::cbuf;

Structs

EventAttrs
Mirrors the C struct for hexchat_event_attrs. It holds the timestamps for the callback invocations for callbacks registered using hexchat_print_attrs(), and similar commands.

Hexchat
This struct mirrors the C Hexchat struct passed to the plugin from Hexchat when the plugin is loaded. Hexchat's API is implemented as a struct holding callbacks to its native functions.

Enums

Eat
The return value for client plugin callbacks.

FD
File descriptor types.

Priority
The priorty for a given callback invoked by Hexchat.

Type Definitions

C_AttrCallback
Mirrors the print attr callback function pointer of Hexchat.

C_Callback
Mirrors the callback function pointer of Hexchat.

C_FDCallback
Mirrors the FD related callback function pointer of Hexchat.

C_PrintCallback
Mirrors the print callback function pointer of Hexchat.

C_TimerCallback
Mirrors the timer callback function pointer of Hexchat.

hexchat_context

hexchat_event_attrs

hexchat_hook

hexchat_list