Module hexchat_api::hexchat_callbacks

This file contains the C-facing functions that are registered directly with Hexchat when a client plugin registers a Rust-facing callback. The callbacks in this file wrap the Rust-facing callbacks, marshal the parameters (word, word_eol, etc) for the Rust callbacks.

Re-exports

use libc::c_int;

use libc::c_char;

use libc::c_void;

use std::ffi::CStr;

use std::panic::catch_unwind;

use crate::callback_data::CallbackData;

use crate::hexchat::Eat;

use crate::hexchat::EventAttrs;

use crate::hexchat_entry_points::HEXCHAT;

use crate::utils::*;

Functions

c_callback
An actual callback registered with Hexchat, which proxies for client plugin callbacks. It builds the String vectors passed to client callbacks. See Hexchat API

c_fd_callback
An actual callback registered with Hexchat, which proxies for client plugin callbacks.

c_print_attrs_callback
An actual callback registered with Hexchat, which proxies for client plugin callbacks. See Hexchat API

c_print_callback
An actual callback registered with Hexchat, which proxies for client plugin callbacks. It builds the String vector and invokes the client plugin's callbacks. The client plugin callback and data is placed within the user_data.

c_timer_callback
An actual callback registered with Hexchat, which proxies for client plugin callbacks.