Module hexchat_api::hexchat_entry_points

This module holds the DLL entry points for this plugin library. These exported functions are what Hexchat links to directly when the plugin is loaded.

These functions register the plugin info with Hexchat and set up to "catch" any panics that Rust-side callbacks might "raise" during execution. Panics are displayed in the currently active Hexchat window/context. The debug build of the library will include a stack trace in the error message.

Re-exports

use backtrace::Backtrace;

use libc::c_int;

use libc::c_char;

use libc::c_void;

use std::ffi::CString;

use std::marker::PhantomPinned;

use std::panic;

use std::panic;

use std::panic::catch_unwind;

use std::panic::UnwindSafe;

use std::pin::Pin;

use std::ptr::null;

use std::ptr::NonNull;

use crate::hexchat::Hexchat;

use crate::utils::*;

Structs

PluginInfo
Holds client plugin information strings.

Statics

HEXCHAT
The global Hexchat pointer obtained from hexchat_plugin_init().

PLUGIN_INFO
Holds persistent client plugin info strings.

Functions

lib_get_info
This function sets Hexchat's character pointer pointer's to point at the pinned buffers holding info about a plugin.

lib_hexchat_plugin_deinit
Invoked indirectly while a plugin is being unloaded. This function will call the deinitialization function that was registered using the dll_entry_points() macro.

lib_hexchat_plugin_get_info
Called indirectly when a plugin is loaded to get info about it. This function calls the client plugin's plugin_get_info() indirectly to obtain the persistent plugin info strings that it sets the paramters to.

lib_hexchat_plugin_init
Called indirectly while a plugin is being loaded. This function will invoke The function that was registered for initialization by the dll_entry_points() macro.

set_panic_hook
Sets the panic hook so panic info will be printed to the active Hexchat window. The debug build includes a stack trace using
Backtrace

Type Definitions

DeinitFn

InfoFn

InitFn
The signatures for the required functions a plugin needs to implement to create a loadable Hexchat plugin.