Module hexchat_api::hook

This object wraps hook pointers returned when callbacks are registered with Hexchat. The Rust-style hook can be used to unhook commands directly via its unhook() function. This object protects against attempts to unhook the same callback more than once, which can crash Hexchat.

The unhook() function returns the user_data that was registered with the associated callback, passing ownership to the caller. Invoking unhook() more than once returns None.

The hooks can be cloned. Internally, clones safely share the same hook pointer. When hooks go out of scope, they do not remove their associated commands. Hooks can be ignored by the plugin if there is no need to unhook commands. The most relevant use of a hook could be to cancel timer callbacks.

Re-exports

use libc::c_void;

use std::any::Any;

use std::cell::RefCell;

use std::rc::Rc;

use std::ptr::null;

use crate::callback_data::*;

use crate::hexchat_entry_points::HEXCHAT;

Structs

Hook
A wrapper for Hexchat callback hooks. These hooks are returned when registering callbacks and can be used to unregister (unhook) them.