Module hexchat_api::list_iterator

This module provides a Rust wrapper for Hexchat's list related API. A list can be accessed by creating a ListIterator by passing the name of one of the list types to the constructor. The iterator itself can be used to access list item fields by name using get_field(). Fields can't be accessed until next() has been invoked to advance the internal pointer. This iterator can be used in loops like any other iterator, or collect() can be called to generate a vector or other collections.

Re-exports

use libc::c_void;

use libc::c_char;

use libc::time_t;

use std::cell::RefCell;

use std::borrow::Borrow;

use std::collections::hash_map::Keys;

use std::collections::HashMap;

use std::error::Error;

use std::fmt;

use std::iter::IntoIterator;

use std::marker::PhantomData;

use std::rc::Rc;

use std::ptr::null;

use std::ffi::CString;

use std::ffi::CStr;

use crate::hexchat::Hexchat;

use crate::hexchat_entry_points::HEXCHAT;

use crate::utils::*;

use FieldValue::*;

use ListError::*;

Structs

ListIterator
The ListIterator wraps the list pointer and related functions of Hexchat. It provides are more Rust OO interface. The iterator returns clones of itself that can be used to access the current list item's fields through get_field(). The list iterator object is internally a smart pointer, among other things. You can clone it if you need multiple references to a list.

ListIteratorData
Holds the iterator state and maps the field names to their data type.

Enums

FieldValue
Field Data Types

ListError
Errors