Struct hexchat_api::list_iterator::ListIteratorData

struct ListIteratorData {
    field_types: Vec<(String, i8)>,
    hc: &'static Hexchat,
    list_ptr: *const c_void,
    started: bool,
}

Holds the iterator state and maps the field names to their data type.

Fields

field_types - A mapping of field names to data type.

field_names - The list of field names for the particular list.

list_ptr - A raw pointer to a list internal to Hexchat.

hc - The Hexchat pointer.

started - true if next() has aready been called on the Rust iter.

Fields
field_types: Vec<(String, i8)>hc: &'static Hexchatlist_ptr: *const c_voidstarted: bool
Implementations

impl ListIteratorData

fn get_type(&self, field: &str) -> Option<i8>

Returns the type of the given field. The field lists are short, so a simple comparisons search for the right item may be quicker than a HashMap's hashings and lookups.