Module hexchat_api::thread_facilities

This module provides facilities for accessing Hexchat from routines
running on threads other than Hexchat's main thread.

Hexchat's plugin API isn't inherently thread-safe, however plugins can spawn separate threads and invoke Hexchat's API by placing routines to execute on Hexchat's main thread.

main_thread() makes it easy to declare a function, or closure, that contains Hexchat API calls. Once executed, it uses the timer feature of Hexchat to delegate. The function or closure can return any sendable cloneable value, and main_thread() will pass that back to the calling thread via an AsyncResult object. This can either be ignored, and the thread can continue doing other work, or AsyncResult.get() can be invoked on the result object; this call will block until the main thread has finished executing the callback.

Re-exports

use std::sync::Arc;

use std::sync::Condvar;

use std::sync::Mutex;

use crate::hexchat::Hexchat;

use crate::hexchat_entry_points::HEXCHAT;

Structs

AsyncResult
A result object that allows callbacks operating on a thread to send their return value to a receiver calling get() from another thread. Whether return data needs to be transferred or not, this object can be used to wait on the completion of a callback, thus providing synchronization between threads.

Functions

main_thread
Executes a closure from the Hexchat main thread. This function returns immediately with an AsyncResult object that can be used to retrieve the result of the operation that will run on the main thread.